hw_13: add relation crud methods for dao

+++ hw_13_crud_hibernate/src/main/java/ua/com/alevel/persistance/dto/GameDto.java
@@ -2,7 +2,7 @@ package ua.com.alevel.persistance.dto;
 
 import ua.com.alevel.persistance.entity.Game;
 
-public record GameDto(Game game, int playerCount) {
+public record GameDto(Game game, long playerCount) {
 
     @Override
     public String toString() {
