+++ hw_13_crud_hibernate/src/main/java/ua/com/alevel/persistance/dto/PlayerDto.java
@@ -2,7 +2,7 @@ package ua.com.alevel.persistance.dto;
 
 import ua.com.alevel.persistance.entity.Player;
 
-public record PlayerDto(Player player, int gamesCount) {
+public record PlayerDto(Player player, long gamesCount) {
 
     @Override
     public String toString() {
